fix: normalize legacy student fee identifier

+++ EduOS.Core/Entities/Finance/StudentFee.cs
@@ -4,7 +4,7 @@ namespace EduOS.Core.Entities.Finance
 {
     public class StudentFee : BaseTenantEntity
     {
-        public int StudentId { get; set; }
+        public long StudentId { get; set; }
         public int FeeTypeId { get; set; }
         public decimal Amount { get; set; }
         public DateTime DueDate { get; set; }
